+++ src/components/Navbar.css
@@ -20,6 +20,24 @@ nav.navbar {
     padding-right: 20px;
     background-color: white;
 }
+
+.btn-outline-primary {
+    --bs-btn-color: #565FA4 !important;
+    --bs-btn-border-color: #565FA4 !important;
+    --bs-btn-hover-color: #fff;
+    --bs-btn-hover-bg: #565FA4 !important;
+    --bs-btn-hover-border-color: #565FA4 !important;
+    --bs-btn-focus-shadow-rgb: 13, 110, 253;
+    --bs-btn-active-color: #fff;
+    --bs-btn-active-bg: #565FA4 !important;
+    --bs-btn-active-border-color: #565FA4 !important;
+    --bs-btn-active-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.125);
+    --bs-btn-disabled-color: #565FA4 !important;
+    --bs-btn-disabled-bg: transparent;
+    --bs-btn-disabled-border-color: #565FA4 !important;
+    --bs-gradient: none;
+}
+
 .custom-icon {
     color: #565FA4 !important;
 }
