+++ src/components/Navbar.jsx
@@ -1,14 +1,54 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import './Navbar.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap-icons/font/bootstrap-icons.css';
 import { Link } from 'react-router-dom';
-import LoadScripts from './Navbarimports.jsx';
 
 const Navbar = () => {
+    const [scriptsLoaded, setScriptsLoaded] = useState(false);
+
+    useEffect(() => {
+        const addScript = (src, integrity, crossOrigin, callback) => {
+            const script = document.createElement('script');
+            script.src = src;
+            script.integrity = integrity;
+            script.crossOrigin = crossOrigin;
+            script.async = true;
+            script.onload = callback;
+            document.body.appendChild(script);
+        };
+
+        const onScriptLoad = () => {
+            console.log('Script zijn geladen en uitgevoerd.');
+            setScriptsLoaded(true);
+        };
+
+        if (!scriptsLoaded) {
+            addScript(
+                'https://code.jquery.com/jquery-3.3.1.slim.min.js',
+                'sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo',
+                'anonymous',
+                () => {
+                    addScript(
+                        'https://cdn.jsdelivr.net/npm/popper.js@1.14.3/dist/umd/popper.min.js',
+                        'sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49',
+                        'anonymous',
+                        () => {
+                            addScript(
+                                'https://cdn.jsdelivr.net/npm/bootstrap@4.1.3/dist/js/bootstrap.min.js',
+                                'sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy',
+                                'anonymous',
+                                onScriptLoad
+                            );
+                        }
+                    );
+                }
+            );
+        }
+    }, [scriptsLoaded]);
+
     return (
         <nav className="navbar navbar-expand-lg border-bottom">
-            <LoadScripts />
             <div className="navbar-container">
                 <button className="btn btn-outline-primary navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <i className="bi bi-list custom-icon"></i>
